pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "robot"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.28.1",
    "python-dotenv>=1.2.2",
    "seleniumbase>=4.47.4",
]

authors = [
    { name = "David Duran", email = "dadch1404@gmail.com" },
]

[project.optional-dependencies]
dev = ["mypy>=1.19.1"]

[project.scripts]
robot = "robot.app.cli:main"

[tool.setuptools.packages.find]
include = ["robot*"]

# ruff config
[tool.ruff]
fix = true
target-version = "py310"

[tool.ruff.lint]
preview = true
extend-select = [
    "A",   # flake8-builtins
    "B",   # flake8-bugbear
    "BLE", # flake8-blind-except
    "C4",  # flake8-comprehensions
    "C90", # mccabe
    "COM", # flake8-commas
    "CPY", # flake8-copyright
    "DTZ", # flake8-datetimez
    "EM",  # flake8-errmsg
    "ERA", # flake8-eradicate/eradicate
    "EXE", # flake8-executable
    "F",   # pyflakes
    "FA",  # flake8-future-annotations
    "FBT", # flake8-boolean-trap
    "FIX", # flake8-fixme
    "FLY", # flynt
    "FURB", # refurb
    "G",   # flake8-logging-format
    "I",   # isort
    "ICN", # flake8-import-conventions
    "INP", # flake8-no-pep420
    "ISC", # flake8-implicit-str-concat
    "LOG", # flake8-logging
    "N",   # pep8-naming
    "PGH", # pygrep
    "PIE", # flake8-pie
    "PTH", # flake8-use-pathlib
    "Q",   # flake8-quotes
    "RET", # flake8-return
    "RSE", # flake8-raise
    "RUF", # ruff checks
    "SIM", # flake8-simplify
    "T10", # flake8-debugger
    # "T20", # flake8-print
    "TC",  # flake8-type-checking
    "TD",  # flake8-todos
    "TID", # flake8-tidy-imports
    "TRY", # tryceratops
    "UP",  # pyupgrade
]

ignore = ["COM812", "TRY300", "CPY001"]

[tool.ruff.lint.flake8-tidy-imports]
ban-relative-imports = "all"

[tool.ruff.lint.isort]
force-single-line = false
lines-between-types = 1
lines-after-imports = 2
order-by-type = true
relative-imports-order = "closest-to-furthest"
extra-standard-library = ["typing"]
section-order = ["future", "standard-library", "third-party", "first-party", "local-folder"]
